// src/@types/Entities.types.ts
export enum EType {
  bug = 'bug',
  dark = 'dark',
  dragon = 'dragon',
  electric = 'electric',
  fairy = 'fairy',
  fire = 'fire',
  fighting = 'fighting',
  flying = 'flying',
  grass = 'grass',
  ghost = 'ghost',
  ground = 'ground',
  ice = 'ice',
  normal = 'normal',
  poison = 'poison',
  psychic = 'psychic',
  rock = 'rock',
  steel = 'steel',
  water = 'water'
}